+++ test/core/dimensions.v
@@ -37,6 +37,10 @@ module top;
 
         $display("args %b", 2);
         $display("args %b", 2);
+        $display("args %b", 1'bx);
+        $display("args %b", 2);
+        $display("args %b", 2);
+        $display("args %b", 1'bx);
         $display("args %b", 2);
         $display("args %b", 2);
         $display("args %b", 2);