+++ test/error/size_cast_neg_lit_1.sv
@@ -1,0 +1,4 @@
+// pattern: size cast width 1'sb1 is not a positive integer
+module top;
+    initial $display((1'sb1)'(2));
+endmodule