+++ test/error/size_cast_neg_lit_2.sv
@@ -1,0 +1,4 @@
+// pattern: size cast width 2'sb11 is not a positive integer
+module top;
+    initial $display((2'sb11)'(2));
+endmodule