+++ test/error/size_cast_neg_var_2.sv
@@ -1,0 +1,5 @@
+// pattern: size cast width 2'sb11 is not a positive integer
+module top;
+    wire x = 0;
+    initial $display((2'sb11)'(x));
+endmodule