+++ test/error/size_cast_x_var.sv
@@ -1,0 +1,5 @@
+// pattern: size cast width 1'bx is not an integer
+module top;
+    wire x = 0;
+    initial $display((1'bx)'(x));
+endmodule