+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+* Non-positive integer size casts are now detected and forbidden
+* Negative indices in struct pattern literals are now detected and forbidden
+
 ## v0.0.8
 
 Future releases will have complete change logs.